fixed cast issue from Long to Integer

+++ src/org/nescent/mmdb/spring/SearchDataController.java
@@ -478,11 +478,11 @@ public class SearchDataController implements Controller {
 	    for (int j = 0; j < objs.length; j++) {
 		String studyName = (String) objs[1];
 		if (studyName.indexOf("inbreed") != -1)
-		    inbreedNum = ((Integer) objs[0]).intValue();
+		    inbreedNum = ((Long) objs[0]).intValue();
 		if (studyName.indexOf("outcross") != -1)
-		    outNum = ((Integer) objs[0]).intValue();
+		    outNum = ((Long) objs[0]).intValue();
 		if (studyName.indexOf("pollination") != -1)
-		    pollNum = ((Integer) objs[0]).intValue();
+		    pollNum = ((Long) objs[0]).intValue();
 	    }
 	}
